file_scanner_mvp/requirements.txt
# Core dependencies
python-magic >= 0.4.27
pytest >= 7.0


# Platform-specific dependencies
windows-curses; sys_platform == 'win32'
